fix: clean up code and add canonical url
